최대 바이트 크기를 100까지 설정할 수 있도록 늘림.

+++ app/src/main/java/com/parkjongeun/mixsimulator/legacy/Assembler.java
@@ -251,7 +251,7 @@ public class Assembler {
                 processFutureRef(memory);
 
                 final Word word = evalWValueW(line.colADDRESS);
-                final int startFrom = word.getQuantity(4, 5);
+                final int startFrom = (int) word.getQuantity(4, 5);
 
             } else {
                 Word word = processMIXOperator(line);
